requirements.txt
numpy>=2.1.2
PySide6>=6.9.0
pyinstaller>=6.8.0
scipy>=1.14.1
pillow>=11.2.1
